# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cruftman",
        "description": "Cruftman",
        "contact": {
            "email": "ptomulik@meil.pw.edu.pl"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/person/{id}": {
            "get": {
                "summary": "Get a person with given ID",
                "description": "Returns a single person",
                "operationId": "getPersonById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of person to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Person not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpException"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/": {
            "get": {
                "summary": "Display welcome message",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Person": {
                "title": "Person model",
                "description": "Person model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uid": {
                        "type": "string"
                    },
                    "givenname": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "birthday": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Person"
                }
            },
            "HttpException": {
                "required": [
                    "status_code",
                    "message"
                ],
                "properties": {
                    "status_code": {
                        "type": "integer",
                        "example": "404"
                    },
                    "message": {
                        "type": "string",
                        "example": "Not found"
                    },
                    "debug": {
                        "$ref": "#/components/schemas/HttpDebug"
                    }
                },
                "type": "object"
            },
            "HttpDebug": {
                "required": [
                    "line",
                    "file",
                    "class",
                    "trace"
                ],
                "properties": {
                    "line": {
                        "type": "integer",
                        "example": 123
                    },
                    "file": {
                        "type": "string",
                        "example": "src/Cruftman/Http/Controllers/PersonController.php"
                    },
                    "class": {
                        "type": "string",
                        "example": "NotFoundHttpException"
                    },
                    "trace": {
                        "$ref": "#/components/schemas/ExceptionDebugTrace"
                    }
                },
                "type": "object"
            },
            "ExceptionDebugTrace": {
                "oneOf": [
                    {
                        "properties": {
                            "previous": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "current": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                ]
            }
        }
    }
}